server/requirements.txt
flask
flask_cors
pandas
requests
beautifulsoup4
langchain-groq
python-dotenv
langchain_community
langchain